refactor[#523] : Region, LatLng 타입 적용
