+++ src/entities/auth/api/getMyInfo.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
+import { Region } from "@/entities/map/@x/auth";
 import { apiClient } from "@/shared/lib";
 import { MY_INFO_END_POINT, SOCIAL_TYPE } from "../constants";
 import { authQueryKey } from "../constants";
-import type { Region } from "./getRegion";
 
 export interface MyInfo {
   email: string;
