+++ src/features/setting/api/postChangeRegion.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { Region } from "@/entities/auth/api";
 import { authQueryKey } from "@/entities/auth/constants";
+import type { Region } from "@/entities/map/types/server";
 import { apiClient } from "@/shared/lib";
 import { SETTING_END_POINT } from "../constants";
