+++ src/features/setting/ui/regionChangeButton.tsx
@@ -1,4 +1,4 @@
-import type { Region } from "@/entities/auth/api";
+import type { Region } from "@/entities/map/types/server";
 import { ActionChip } from "@/shared/ui/chip";
 import { ArrowRightIcon } from "@/shared/ui/icon";
 import { useChangeRegionModal } from "../hooks";
